Move TagHelperTooltipFactories to Workspaces layer
