+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Resources/SR.resx
@@ -141,17 +141,8 @@
   <data name="Generate_Event_Handler_Title" xml:space="preserve">
     <value>Generate Event Handler '{0}'</value>
   </data>
-  <data name="Not_Available_In" xml:space="preserve">
-    <value>Not available in</value>
-  </data>
   <data name="ReTrigger_Completions_Title" xml:space="preserve">
     <value>"Re-trigger completions..."</value>
-  </data>
-  <data name="TagHelper_Attribute_Glyph" xml:space="preserve">
-    <value>Razor TagHelper Attribute Glyph</value>
-  </data>
-  <data name="TagHelper_Element_Glyph" xml:space="preserve">
-    <value>Razor TagHelper Element Glyph</value>
   </data>
   <data name="Unknown_ProjectChangeKind" xml:space="preserve">
     <value>Unknown ProjectChangeKind {0}</value>
