+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Resources/xlf/SR.de.xlf
@@ -42,11 +42,6 @@
         <target state="translated">Ereignishandler "{0}" generieren</target>
         <note />
       </trans-unit>
-      <trans-unit id="Not_Available_In">
-        <source>Not available in</source>
-        <target state="translated">Nicht verfügbar in</target>
-        <note />
-      </trans-unit>
       <trans-unit id="ReTrigger_Completions_Title">
         <source>"Re-trigger completions..."</source>
         <target state="translated">"Abschlüsse erneut auslösen..."</target>
@@ -55,16 +50,6 @@
       <trans-unit id="Statement">
         <source>statement</source>
         <target state="translated">Anweisung</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="TagHelper_Attribute_Glyph">
-        <source>Razor TagHelper Attribute Glyph</source>
-        <target state="translated">Razor TagHelper-Attributsymbol</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="TagHelper_Element_Glyph">
-        <source>Razor TagHelper Element Glyph</source>
-        <target state="translated">Razor TagHelper-Elementsymbol</target>
         <note />
       </trans-unit>
       <trans-unit id="Unknown_ProjectChangeKind">
