+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Resources/xlf/SR.ko.xlf
@@ -42,11 +42,6 @@
         <target state="translated">이벤트 처리기 '{0}' 생성</target>
         <note />
       </trans-unit>
-      <trans-unit id="Not_Available_In">
-        <source>Not available in</source>
-        <target state="translated">에서 사용할 수 없음</target>
-        <note />
-      </trans-unit>
       <trans-unit id="ReTrigger_Completions_Title">
         <source>"Re-trigger completions..."</source>
         <target state="translated">"완료된 항목 다시 트리거"</target>
@@ -55,16 +50,6 @@
       <trans-unit id="Statement">
         <source>statement</source>
         <target state="translated">문</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="TagHelper_Attribute_Glyph">
-        <source>Razor TagHelper Attribute Glyph</source>
-        <target state="translated">Razor TagHelper 특성 문자 모양</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="TagHelper_Element_Glyph">
-        <source>Razor TagHelper Element Glyph</source>
-        <target state="translated">Razor TagHelper 요소 문자 모양</target>
         <note />
       </trans-unit>
       <trans-unit id="Unknown_ProjectChangeKind">
