+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Resources/SR.resx
@@ -163,4 +163,13 @@
   <data name="Language_Services_Missing_Service" xml:space="preserve">
     <value>Razor language services not configured properly, missing language service '{0}'.</value>
   </data>
+  <data name="Not_Available_In" xml:space="preserve">
+    <value>Not available in</value>
+  </data>
+  <data name="TagHelper_Attribute_Glyph" xml:space="preserve">
+    <value>Razor TagHelper Attribute Glyph</value>
+  </data>
+  <data name="TagHelper_Element_Glyph" xml:space="preserve">
+    <value>Razor TagHelper Element Glyph</value>
+  </data>
 </root>
